@@ -1,7 +1,6 @@
 [build-system]
 requires = [
-    "setuptools < 45 ; python_version < '3'",
-    "setuptools ; python_version > '2.7'",
+    "setuptools",
     "wheel",
 ]
 build-backend = "setuptools.build_meta"
@@ -26,8 +25,6 @@ classifiers = [
     "Programming Language :: Python",
     "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
     "Topic :: Software Development :: Libraries :: Python Modules",
-    "Programming Language :: Python :: 2",
-    "Programming Language :: Python :: 2.7",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
@@ -37,14 +34,13 @@ classifiers = [
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
 ]
-requires-python = ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*, !=3.5.*, !=3.6.*"
+requires-python = ">=3.7"
 dependencies = [
     "Flask > 1.0",
     "Jinja2 >= 2.10.2",
     "Markdown >= 2.5",
     "PyYAML > 5.3.1",
     "pytz; python_version=='2.7'",
-    "six",
 ]
 [project.urls]
 Repository = "https://github.com/Flask-FlatPages/Flask-FlatPages"
@@ -68,8 +64,6 @@ envlist = [
     "docs",
     "lint",
     "fmt",
-    "py27",
-    "py27-pygments",
     "py37",
     "py37-pygments",
     "py38",
@@ -101,9 +95,6 @@ commands = [
         "{posargs}",
     ],
 ]
-
-[tool.tox.env.py27-pygments]
-extras = ["codehilite"]
 
 [tool.tox.env.py37-pygments]
 extras = ["codehilite"]
