@@ -1,15 +1,11 @@
 """Utility functions to render Markdown text to HTML."""
 
 import markdown
-import six
 from markdown.extensions import codehilite
 
 from .imports import PygmentsHtmlFormatter
 
-if six.PY3:
-    from io import StringIO
-else:
-    from StringIO import StringIO
+from io import StringIO
 
 
 class NamedStringIO(StringIO, object):
@@ -23,16 +19,13 @@ class NamedStringIO(StringIO, object):
         :param name: The name to attach to the stream. Will
             be consumed in e.g. ReaderErrors raised by pyyaml.
         """
-        if not six.PY3:
-            super(NamedStringIO, self).__init__(content)
-        else:
-            super().__init__(content)
+        super().__init__(content)
         self.name = name
 
 
 def force_unicode(value, encoding="utf-8", errors="strict"):
     """Convert bytes or any other Python instance to string."""
-    if isinstance(value, six.text_type):
+    if isinstance(value, str):
         return value
     return value.decode(encoding, errors)
 
@@ -68,14 +61,14 @@ def pygmented_markdown(text, flatpages=None):
 
         for extension in original_extensions:
             if (
-                isinstance(extension, six.string_types)
+                isinstance(extension, str)
                 and "codehilite" in extension
             ):
                 continue
             elif isinstance(extension, codehilite.CodeHiliteExtension):
                 continue
             extensions.append(extension)
-            if isinstance(extension, six.string_types):
+            if isinstance(extension, str):
                 if extension in original_config:
                     extension_configs[extension] = original_config[extension]
     elif not extensions:
